refactor: separate reactive header logic to custom hook
반응형 헤더 관련 로직을 `useReactive()` 커스텀 훅으로 분리함.
